Removes Handle from api endpoint names
The word Handle on the end of a endpoint names was a holdover from Ardalis's API Endpoints base classes implementing a `Handle` method.

I left the WithName methods in place since the auto generated operation names have the verb and version (e.g. "GetApiV1Todolists").  It may be worth a discussion about the naming conventions and looking into a way to auto generate them.